Extract applyCodeFocus into shared code-focus.js module
The implementation was duplicated between application.js and export.js
with subtle differences (rAF wrapper vs direct await, forEach vs for-of).

Extract the canonical async for-of implementation to code-focus.js.
Both callers import it:
- application.js: fire-and-forget (same behaviour as before)
- export.js: awaited (needed before signalling presently:ready)
